feat: disabled dnd
#refs WEB20-285

+++ src/app/shared/components/gridtree/gridtree.component.ts
@@ -20,6 +20,7 @@ import { GridTreeService } from './gridtree.service';
 })
 export class GridTreeComponent<T> implements OnDestroy {
   @Input() columns: Array<IGridTreeColumn<T>> = [];
+  @Input() dnd = false;
 
   @Output() select = this.gridTreeService.select.map(row => row.data);
   @Output() dblclick = this.gridTreeService.dblclick.map(row => row.data);
@@ -32,25 +33,29 @@ export class GridTreeComponent<T> implements OnDestroy {
     private cdRef: ChangeDetectorRef,
     private gridTreeService: GridTreeService<T>,
   ) {
-    this.gridTreeServiceSubscription = this.gridTreeService.drop.subscribe((event: IGridTreeDragAndDropEvent<T>) => {
-      if (
-        this.idGetter(event.draggedRow) === this.idGetter(event.targetRow) ||
-        this.gridTreeService.isChild(event.targetRow, event.draggedRow, this.idGetter)
-      ) {
-        return;
-      }
-      this._rows = this.gridTreeService.removeRowFrom(this._rows, event.draggedRow, this.idGetter);
-      if (event.type === GridTreeDragAndDropEventTypeEnum.INTO) {
-        this._rows = this.gridTreeService.addRowTo(this._rows, event.draggedRow, event.targetRow, this.idGetter);
-      } else {
-        this._rows = this.gridTreeService.addRowAfter(this._rows, event.draggedRow, event.targetRow, this.idGetter);
-      }
-      this.cdRef.markForCheck();
-    });
+    if (this.dnd) {
+      this.gridTreeServiceSubscription = this.gridTreeService.drop.subscribe((event: IGridTreeDragAndDropEvent<T>) => {
+        if (
+          this.idGetter(event.draggedRow) === this.idGetter(event.targetRow) ||
+          this.gridTreeService.isChild(event.targetRow, event.draggedRow, this.idGetter)
+        ) {
+          return;
+        }
+        this._rows = this.gridTreeService.removeRowFrom(this._rows, event.draggedRow, this.idGetter);
+        if (event.type === GridTreeDragAndDropEventTypeEnum.INTO) {
+          this._rows = this.gridTreeService.addRowTo(this._rows, event.draggedRow, event.targetRow, this.idGetter);
+        } else {
+          this._rows = this.gridTreeService.addRowAfter(this._rows, event.draggedRow, event.targetRow, this.idGetter);
+        }
+        this.cdRef.markForCheck();
+      });
+    }
   }
 
   ngOnDestroy(): void {
-    this.gridTreeServiceSubscription.unsubscribe();
+    if (this.dnd) {
+      this.gridTreeServiceSubscription.unsubscribe();
+    }
   }
 
   get rows(): Array<IGridTreeRow<T>> {
